decouple SDK utilities from matchmaking code
