+++ EdgegapSDK.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Edgegap.Matchmaking.SDK",
+    "name": "Edgegap.SDK",
     "rootNamespace": "",
     "references": [],
     "includePlatforms": [],
