+++ Runtime/Logger.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace Edgegap.Matchmaking
+namespace Edgegap
 {
     public static class Logger
     {
@@ -9,29 +9,35 @@ namespace Edgegap.Matchmaking
         {
             if (!Debug.isDebugBuild)
                 return;
-            Debug.Log(_FormatLog(message));
+            Debug.Log(message);
         }
 
         public static void _Warn<T>(T message)
         {
-            Debug.LogWarning(_FormatLog(message));
+            Debug.LogWarning(message);
         }
 
         public static void _Error<T>(T message)
         {
-            Debug.LogError(_FormatLog(message));
+            Debug.LogError(message);
         }
 
-        public static string _FormatLog<T>(T message) =>
-            $"{DateTime.UtcNow} Matchmaker | {_ToStringOrNull(message)}";
-
-        public static string _FormatNotifyMessage<T>(string prefix, string message, T value)
+        public static string _FormatNotifyMessage<T>(
+            string service,
+            string subject,
+            string message,
+            T value
+        )
         {
-            return $"{prefix}.notify / {message}\n{_ToStringOrNull(value)}";
+            return string.Join(
+                "\n",
+                new string[] { $"{service} | {subject}.notify / {message}", _ToStringOrNull(value) }
+            );
         }
 
         public static string _FormatUpdateMessage<T>(
-            string prefix,
+            string service,
+            string subject,
             string message,
             T previous,
             T current
@@ -41,12 +47,11 @@ namespace Edgegap.Matchmaking
                 "\n",
                 new string[]
                 {
-                    $"{prefix}.changed / {message}",
+                    $"{service} | {subject}.changed / {message}",
                     $"current: {_ToStringOrNull(current)}",
                     $"previous: {_ToStringOrNull(previous)}",
                 }
             );
-            ;
         }
 
         public static string _ToStringOrNull<T>(T value)
@@ -55,4 +60,3 @@ namespace Edgegap.Matchmaking
         }
     }
 }
-
