+++ Runtime/Matchmaking/DTOs/BeaconsResponseDTO.cs.meta
