+++ Runtime/Ping.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 7d3d83374c8a859418095b37c8fe25d1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
