+++ Runtime/SafeHttpRequest.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 713b6d9e5ff51214182579a4d76f4bf9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
