+++ package.json
@@ -1,27 +1,27 @@
 {
-    "name": "com.edgegap.unity-matchmaking-sdk",
-    "version": "2.1.0",
-    "displayName": "Edgegap Matchmaking SDK",
-    "description": "Get started quickly with Edgegap Matchmaking Matchmaking.",
-    "unity": "2021.3",
-    "unityRelease": "0f1",
-    "documentationUrl": "https://docs.edgegap.com/learn/matchmaking/getting-started",
-    "hideInEditor": false,
-    "changelogUrl": "https://docs.edgegap.com/learn/matchmaking/matchmaker-in-depth#changelog",
-    "licensesUrl": "https://choosealicense.com/licenses/mit/",
-    "samples": [
-        {
-            "displayName": "Simple Example",
-            "description": "Minimal Recommended Matchmaking Example.",
-            "path": "Samples~/SimpleExample"
-        }
-    ],
-    "dependencies": {
-        "com.unity.nuget.newtonsoft-json": "3.2.1"
-    },
-    "author": {
-        "name": "Edgegap Technologies",
-        "email": "contact@edgegap.com",
-        "url": "https://www.edgegap.com"
+  "name": "com.edgegap.unity-sdk",
+  "version": "3.0.0",
+  "displayName": "Edgegap SDK",
+  "description": "Get started quickly with Edgegap Matchmaking and Server Browser.",
+  "unity": "2021.3",
+  "unityRelease": "0f1",
+  "documentationUrl": "https://docs.edgegap.com/unity/developer-tools#sdk",
+  "hideInEditor": false,
+  "changelogUrl": "https://github.com/edgegap/edgegap-unity-sdk/releases",
+  "licensesUrl": "https://choosealicense.com/licenses/mit/",
+  "samples": [
+    {
+      "displayName": "Matchmaking - Simple Example",
+      "description": "Minimal Recommended Example.",
+      "path": "Samples~/MatchmakingSimpleExample"
     }
+  ],
+  "dependencies": {
+    "com.unity.nuget.newtonsoft-json": "3.2.2"
+  },
+  "author": {
+    "name": "Edgegap Technologies",
+    "email": "contact@edgegap.com",
+    "url": "https://www.edgegap.com"
+  }
 }
